fixed collision, fixed damage bounds thing

+++ BallAndPaddle Importer/src/org/alia4j/languages/bp/action/IntAttributeBoundsAssuranceAction.java
@@ -11,7 +11,7 @@ public class IntAttributeBoundsAssuranceAction {
 		TypeHierarchyProvider.findOrCreateFromClass(IntAttributeBoundsAssuranceAction.class),
 		"apply",
 		TypeHierarchyProvider.findOrCreateFromNormalTypeNames(new String[] {"int","int"}),
-		TypeHierarchyProvider.findOrCreateFromClass(double.class),
+		TypeHierarchyProvider.findOrCreateFromClass(int.class),
 		ResolutionStrategy.STATIC
 	);
